Reintroduces support for python 3.2.
Python 3.2 has some changes that break our bootstrapping code, and require that more
of the stdlib is kept outside of the stdlib. To work around this behavior we
now use Py_SetPath to ensure the zipfile is after the plain directory in sys.path
and extract bits of the encodings package (and the codecs module), the bootstrap
code then restores the full encodings package.

Fixes #9
Fixes #6