Fix for issue #32: app bundle crashes on close when using PySide
PySide installs a number of shared libraries into /usr/lib and those
files don't get included by macho_standalone by default. This patch
adds a recipe that tells macholib to act as if those files are not
system files even though they are in a system location.

This requires a very recent version of macholib (not yet released)